requirements.txt
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.0
chromadb>=0.4.22
openai>=1.12.0
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-anthropic>=0.1.0
anthropic>=0.18.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
rich>=13.7.0
sentence-transformers>=2.2.0
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# MCP Server
mcp>=1.0.0

# Slack Bot
slack-bolt>=1.18.0
aiohttp>=3.9.0
